requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
redis==5.0.1
celery==5.3.4
openai==1.3.9
supabase==2.3.5
stripe==7.4.0
fpdf2==2.7.4
